WIP: Add initial cuda_versions capabilities
Allow for build_env.py to generate variants for multiple cuda versions.

The flag --cuda_versions has been added to the default build_env.py script.

Each cuda_version will generate a new variant and build with that version.

In addition, a check is made to ensure the server building the frameworks, is
capable of running 11.0 cuda images.